+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   [#278](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/278)
 * Upgrading to Flask 3.x
   [#394](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/394)
+* Item properties for ISBN identifier
+  [#169](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/169)
 
 ### Fixed